Rewrite of oauth.webapp
The original API was not sufficient. It was not able to adapt to the different needs web applications may have. With OAuth you may not always want to stricly enforce user login. And also, you may want to support multiple OAuth providers in a single app. The original API did not handle that very nicely.

- Removed: constructor (default construction is sufficient now)
- Replaced: performOAuth() by isLoggedIn() and login()
- Removed: unauthenticated()

Application developers now need to write their own performOAuth() like handler to check user login status. This would be a wrapper around isLoggedIn(), with some application specific handling of the user's login status. (So that is in fact a replacement for unauthenticated(), but more flexible.)